[Move/#45] 도메인 모델로 변경
