--- app/src/main/java/com/poti/android/presentation/party/create/model/MemberPriceOption.kt
@@ -1,7 +1,0 @@
-package com.poti.android.presentation.party.create.model
-
-data class MemberPriceOption(
-    val memberId: Long,
-    val name: String,
-    val price: String,
-)
